# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/properties": {
      "post": {
        "operationId": "CreatePropertyController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyRequestDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new property",
        "tags": [
          "Properties"
        ]
      },
      "get": {
        "operationId": "GetPropertiesController_handle",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "VACANT",
                "LISTED",
                "OCCUPIED",
                "MAINTENANCE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of properties"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List properties for the authenticated manager",
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/properties/{id}/status": {
      "patch": {
        "operationId": "UpdatePropertyStatusController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Property UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "400": {
            "description": "Invalid status transition"
          },
          "404": {
            "description": "Property not found"
          },
          "409": {
            "description": "Property has active lease"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update property status",
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/expenses": {
      "post": {
        "operationId": "CreateExpenseController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "propertyId",
                  "category",
                  "amount",
                  "description",
                  "expenseDate"
                ],
                "properties": {
                  "propertyId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "category": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "expenseDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "maintenanceRequestId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense created"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new expense",
        "tags": [
          "Expenses"
        ]
      },
      "get": {
        "operationId": "GetExpensesController_handle",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "propertyId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Expenses list"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get expenses with filters",
        "tags": [
          "Expenses"
        ]
      }
    },
...
    "/expenses/{id}": {
      "get": {
        "operationId": "GetExpenseByIdController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Expense UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense found"
          },
          "404": {
            "description": "Expense not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get expense by ID",
        "tags": [
          "Expenses"
        ]
      },
      "put": {
        "operationId": "UpdateExpenseController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Expense UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "expenseDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "maintenanceRequestId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense updated"
          },
          "404": {
            "description": "Expense not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an expense",
        "tags": [
          "Expenses"
        ]
      },
      "delete": {
        "operationId": "DeleteExpenseController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Expense UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Expense deleted"
          },
          "404": {
            "description": "Expense not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an expense",
        "tags": [
          "Expenses"
        ]
      }
    },
...
    "/vendors": {
      "post": {
        "operationId": "CreateVendorController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "specialty"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "specialty": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vendor created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new vendor",
        "tags": [
          "Vendors"
        ]
      },
      "get": {
        "operationId": "GetVendorsController_handle",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "specialty",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Vendors list"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get vendors",
        "tags": [
          "Vendors"
        ]
      }
    },
    "/vendors/{id}": {
      "get": {
        "operationId": "GetVendorByIdController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vendor UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendor found"
          },
          "404": {
            "description": "Vendor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get vendor by ID",
        "tags": [
          "Vendors"
        ]
      },
      "put": {
        "operationId": "UpdateVendorController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vendor UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "specialty": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string",
                    "nullable": true
                  },
                  "email": {
                    "type": "string",
                    "nullable": true
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vendor updated"
          },
          "404": {
            "description": "Vendor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a vendor",
        "tags": [
          "Vendors"
        ]
      },
      "delete": {
        "operationId": "DeleteVendorController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vendor UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Vendor deleted"
          },
          "404": {
            "description": "Vendor not found"
          },
          "409": {
            "description": "Vendor has open maintenance requests"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a vendor",
        "tags": [
          "Vendors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mobile API",
    "description": "Auto-generated API docs with Swagger",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePropertyRequestDTO": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "123 Main St"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "zipCode": {
            "type": "string",
            "example": "10001"
          },
          "propertyType": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "CONDO",
              "TOWNHOUSE",
              "STUDIO"
            ],
            "example": "APARTMENT"
          },
          "bedrooms": {
            "type": "number",
            "example": 2
          },
          "bathrooms": {
            "type": "number",
            "example": 1.5
          },
          "sqft": {
            "type": "number",
            "example": 1200
          },
          "rentAmount": {
            "type": "number",
            "example": 2500
          },
          "description": {
            "type": "string",
            "example": "Spacious 2BR apartment with city views"
          }
        },
        "required": [
          "address",
          "city",
          "state",
          "zipCode",
          "propertyType",
          "bedrooms",
          "bathrooms",
          "rentAmount"
        ]
      },
...
    }
  }
}